+++ mail/mail-html-transformer/src/sanitizer.rs
@@ -104,7 +104,6 @@ static TAG_SET: LazyLock<HashSet<&'static str>> = LazyLock::new(|| {
         "span",
         "strike",
         "strong",
-        "style",
         "sub",
         "summary",
         "sup",
@@ -251,7 +250,8 @@ static ATTR_SET: LazyLock<HashSet<&'static str>> = LazyLock::new(|| {
 /// Tags that should be removed with their inner HTML.
 static TAGS_TO_REMOVE_WITH_INNER_HTML: LazyLock<HashSet<&'static str>> = LazyLock::new(|| {
     hash_set! {
-        "script"
+        "script",
+        "style",
     }
 });